requirements.txt
flask
mysql-connector-python
streamlit
pandas
python-dotenv
matplotlib
fetch